Refactor tests to split them among multiple subclasses
